Make async example into a test, and commonize test infra.
